requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pdf2image
Pillow
PyMuPDF

